+++ api/src/main/java/com/uade/api/models/ReclamoModel.java
@@ -4,6 +4,9 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigInteger;
+
 @Entity
 @Table(name = "reclamos")
 @Data
@@ -12,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class ReclamoModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idReclamo;
+    private int idReclamo;
     private String estado;
     private String descripcion;
     @OneToOne
