+++ api/src/main/java/com/uade/api/models/UnidadModel.java
@@ -4,6 +4,9 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigInteger;
+
 @Entity
 @Table(name = "unidades")
 @Data
@@ -12,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class UnidadModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idUnidad;
+    private int idUnidad;
     private int piso;
     private int numero;
     private boolean habitada;
